+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Set-SEPCloudRegion function to set the correct region for the SEPCloud API
 - Update-AllowListPolicyByFileName function to add a filename to the allow list policy
+- Update-AllowListPolicyByFileHash function to add a file hash to the allow list policy
 
 
 ### Changed